refactor(serve): split monolithic serving gateway into modular APIRouters
- Extracted REST and WebSocket routes into routers/market.py, routers/retrieval.py, and routers/services.py
- Refactored dependencies to request.app and websocket.app contexts to prevent circular imports
- Established a shared websocket_manager singleton in websocket.py
- Updated documentation (serve README.md) and memory files
